backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pytesseract>=0.3.10
Pillow>=10.2.0
pydantic>=2.5.0
httpx>=0.25.2
elevenlabs>=0.2.26
